% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basil.R
\docType{data}
\name{basil}
\alias{basil}
\title{Basil Dataset}
\format{
A data frame with n rows and m variables:
\describe{
  \item{ID}{ }
  \item{variety}{ }
  \item{sowing_date}{ }
  \item{repot_date}{ }
  \item{harvest_date}{ }
  \item{level}{ }
  \item{lightcolour}{ in Kelvin}
  \item{lightduration}{ in hours}
  \item{darknessduration}{ in hours}
  \item{potspacing}{ in mm}
  \item{row}{ }
  \item{column}{ }
  \item{weight_total}{ in g}
  \item{weight_plant1}{ in g}
  \item{weight_plant2}{ in g}
  \item{weight_plant3}{ in g}
  \item{weight_plant4}{ in g}
  \item{weight_plant5}{ in g}
  \item{weight_plant6}{ in g}
  \item{weight_plant7}{ in g}
  \item{weight_plant8}{ in g}
  \item{weight_plant9}{ in g}
  \item{weight_plant10}{ in g}
  \item{weight_plant11}{ in g}
  \item{weight_plant12}{ in g}
}
}
\source{
\url{https://www.bchwtz.de}

\url{https://github.com/bchwtz/bchwtz-farm}
}
\usage{
basil
}
\description{
Basil
}
\details{
The variables [pop] and [gdp] are geometric means, which is a rather interesting choice.
Both indicators aim at measuring the size of the connection between the two cities, either
in population or monetary means. As the arithmetic mean is heavily influenced by outliers
a connection travelling from a rather small to a large city may not be described accurately
by e.g. the arithmetic mean of the populations. The same may be true for the difference between
both populations. Therefore the geometric mean is chosen by the authors of the data set.
}
\keyword{datasets}
