% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityenergyconsumption.R
\docType{data}
\name{cityenergyconsumption}
\alias{cityenergyconsumption}
\title{City Energy Consumption Data}
\format{
A data frame with 52416 rows and 7 variables:
\describe{
  \item{ date }{ }
  \item{ temperature }{ }
  \item{ humidity }{ }
  \item{ windspeed }{ }
  \item{ generaldiffuseflows }{ }
  \item{ diffuseflows }{}
  \item{ consumption }{Zone 1 Power consumption in UNIT }
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/849/power+consumption+of+tetouan+city}
}
\usage{
cityenergyconsumption
}
\description{
Description TODIO
}
\references{
Salam, A., & El Hibaoui, A. (2018, December). Comparison of Machine Learning Algorithms for the Power Consumption Prediction:-Case Study of Tetouan city“. In 2018 6th International Renewable and Sustainable Energy Conference (IRSEC) (pp. 1-5). IEEE.
}
\keyword{datasets}
