% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhswf.R, R/render.R
\name{exercises}
\alias{exercises}
\alias{render_exercises}
\title{Exercise Sheets PDF Themes}
\usage{
exercises(...)

render_exercises(...)
}
\arguments{
\item{...}{divers parameters}
}
\description{
Exercise Sheets PDF Themes
}
\seealso{
Three other packages also offer RMarkdown interfaces to LaTeX
(or HTML) content: \code{\link[tint]{tint}} for Tufte-style pdf and html,
}
\author{
Stefan Böcker
}
