% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhswf-package.R
\docType{package}
\name{fhswf-package}
\alias{fhswf}
\alias{fhswf-package}
\title{fhswf: FH SWF Package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stefan Böcker \email{boecker.stefan@fh-swf.de} (\href{https://orcid.org/0000-0001-8714-6261}{ORCID})

}
\keyword{internal}
