% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footballsmall.R
\docType{data}
\name{footballsmall}
\alias{footballsmall}
\title{Small Football Dataset}
\format{
A data frame with 18 rows and 14 variables:
\describe{
  \item{club}{Participating teams in the 1. Bundesliga}
  \item{clubabbr}{TODO}
  \item{points}{Sum of points at the end of the season (win=3, draw=1, loss=0)}
  \item{goals_scored}{Total number of scored goals of each team in the season}
  \item{goals_received}{Total number of conceded goals of each team in the season}
  \item{marketvalue}{Monetary worth of a team at the end of the season in million Euros}
  \item{marketvalue_goal}{TODO}
  \item{marketvalue_a}{TODO}
  \item{marketvalue_b}{TODO}
  \item{marketvalue_c}{TODO}
  \item{transferspending}{Total spending used to buy new players by the respective club for the next season}
  \item{age_average}{Average age of the players inside a squad in each team in the season}
  \item{viewers}{Average number of spectators at home matches in the season}
  \item{season}{Football season of the observation}
}
}
\source{
\url{https://www.transfermarkt.de/TODO:WODENNGENAU}
}
\usage{
footballsmall
}
\description{
TODO: Adriane Sommer & Marc Hoffmann
}
\keyword{datasets}
