% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Medical Cost Personal Dataset}
\format{
A data frame with 1338 rows and 7 variables:
\describe{
  \item{charges}{Individual yearly medical costs billed by the health insurance. Unit in US dollars.}
  \item{age}{Age in years of the primary beneficiary.}
  \item{sex}{Insurance contractor gender, female, male}
  \item{bmi}{Body Mass Index, providing an understanding of body, weights that are relatively high or low relative to height. Calculates as ratio of height to weight (kg/m^2). Normal Weight is considered to be in the range from 18.5 to 24.9.}
  \item{children}{Number of children/dependents covered by health insurance.}
  \item{smoker}{Indicates if primary beneficirary is a smoker.}
  \item{region}{The beneficiary's residential area in the US: northeast, southeast, southwest, northwest.}
}
}
\source{
\url{https://github.com/stedy/Machine-Learning-with-R-datasets#readme}

\url{https://www.kaggle.com/datasets/mirichoi0218/insurance}
}
\usage{
insurance
}
\description{
This dataset contains 1,339 medical insurance records. The individual medical
costs billed by health insurance are the variable charges, and the rest of
columns contain personal information such as age, gender, family status, and
whether the patient smokes among other features.
}
\references{
Lantz, Brett, "Machine Learning with R", Third Edition, 2019, Packt Publishing.
}
\keyword{datasets}
