% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Modal Value}
\usage{
modal(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numbers , characters, logicals or factors}

\item{na.rm}{logical. If TRUE, NA values are ignored. If FALSE, NA is returned if x has any NA values}
}
\description{
The mode, or modal value, is the most frequent value in a set of values.
}
