% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistik.R
\name{sd_pop}
\alias{sd_pop}
\title{Statistics function used at FH SWF within statistics course}
\usage{
sd_pop(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object but not a factor coercible to numeric by as.double(x).}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Statistics function used at FH SWF within statistics course
}
\author{
Prof. Dr. Stefan Böcker
}
