% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhswf.R, R/render.R
\name{seminarpaper}
\alias{seminarpaper}
\alias{render_seminarpaper}
\title{Seminar Paper PDF Themes}
\usage{
seminarpaper(...)

render_seminarpaper(...)
}
\description{
Seminar Paper PDF Themes
}
\seealso{
Three other packages also offer RMarkdown interfaces to LaTeX
(or HTML) content: \code{\link[tint]{tint}} for Tufte-style pdf and html,
}
\author{
Stefan Böcker
}
