% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statData.R
\docType{data}
\name{stat2022}
\alias{stat2022}
\title{Data for Statistics 2022}
\format{
A list with data
}
\source{
\url{https://www.bchwtz.de/}
}
\usage{
stat2022
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
