% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whodata.R
\docType{data}
\name{whodata}
\alias{whodata}
\title{WHO Data}
\format{

}
\source{
\url{http://ec2-54-174-131-205.compute-1.amazonaws.com/API/Information.php}
}
\usage{
whodata
}
\description{
A dataset containing information about the development of (almost) all
countries on earth from 1990 to 2019 (excluding some years).

The data was collected using the Human Development Report Office Statistical
Data API (HDRO).
}
\keyword{datasets}
