% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytsubs.R
\docType{data}
\name{ytsubs.bibi}
\alias{ytsubs.bibi}
\title{Youtube Subscribers for Channel "Bibis Beauty Palace"}
\format{
A data frame with 1094 rows and 2 variables:
\describe{
  \item{date}{The date of the observation}
  \item{TotalSubscribers}{Total Subscriber count on Youtube}
}
}
\source{
\url{https://socialblade.com/youtube/user/bibisbeautypalace}

\url{https://www.youtube.com/channel/UCHfdTAyg5t4mb1G-3rJ6QsQ}
}
\usage{
ytsubs.bibi
}
\description{
A dataset containing the total Youtube subscribers for the channel "Bibis Beauty Palace".
}
\keyword{datasets}
