% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytsubs.R
\docType{data}
\name{ytsubs.tseries}
\alias{ytsubs.tseries}
\title{Youtube Subscribers for Channel "T-Series"}
\format{
A data frame with 1094 rows and 2 variables:
\describe{
  \item{date}{The date of the observation}
  \item{TotalSubscribers}{Total Subscriber count on Youtube}
}
}
\source{
\url{https://socialblade.com/youtube/user/tseries}

\url{https://www.youtube.com/channel/UCq-Fj5jknLsUf-MWSy4_brA}
}
\usage{
ytsubs.tseries
}
\description{
A dataset containing the total Youtube subscribers for the channel "T-Series".
}
\keyword{datasets}
