% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basil.R
\docType{data}
\name{basil}
\alias{basil}
\title{Basil Dataset}
\format{
A data frame with n rows and m variables:
\describe{
  \item{ID}{ }
  \item{variety}{ }
  \item{sowing_date}{ }
  \item{repot_date}{ }
  \item{harvest_date}{ }
  \item{level}{ }
  \item{plantplatepos}{ }
  \item{lightcolour}{ }
  \item{lightduration}{ }
  \item{darknessduration}{ }
  \item{potspacing}{ }
  \item{row}{ }
  \item{column}{ }
  \item{plantid}{ }
  \item{comment}{ }
  \item{weight_total}{ Total weight of all plants in a pot in grams. Result stems from a single measurement, directly after harvesting the plants. }
  \item{weight_plant1}{ }
  \item{weight_plant2}{ }
  \item{weight_plant3}{ }
  \item{weight_plant4}{ }
  \item{weight_plant5}{ }
  \item{weight_plant6}{ }
  \item{weight_plant7}{ }
  \item{weight_plant8}{ }
  \item{weight_plant9}{ }
  \item{weight_plant10}{ }
  \item{weight_plant11}{ }
  \item{weight_plant12}{ }
  \item{plantcount}{ }
  \item{weight_plantsum}{ }
  \item{plantdensity}{ }
  \item{weight_lostyield}{ }
  \item{mingroovedist}{ }
  \item{border_window}{ }
  \item{border_door}{ }
  \item{border_water}{ }
  \item{border_elec}{ }
  \item{border_row}{ }
  \item{border_plantplate}{ }
  \item{border_spaceplate}{ }
  \item{border}{Dummy indicating wether a position on the plant plate is located at a border.}
}
}
\source{
\url{https://www.bchwtz.de}

\url{https://github.com/bchwtz/bchwtz-farm}
}
\usage{
basil
}
\description{
Basil
}
\details{
vertical farm
}
\keyword{datasets}
