% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carprices.R
\docType{data}
\name{carprices}
\alias{carprices}
\title{Car Price Dataset}
\format{
A data frame with 205 rows and 26 variables:
\describe{
  \item{body}{}
  \item{brand}{}
  \item{doors}{}
  \item{drive_wheels}{}
  \item{engine_aspiration}{}
  \item{engine_bore}{}
  \item{engine_compression}{}
  \item{engine_cylinders}{}
  \item{engine_fuelsystem}{}
  \item{engine_fueltype}{}
  \item{engine_location}{}
  \item{engine_size}{}
  \item{engine_stroke}{}
  \item{engine_type}{}
  \item{height}{}
  \item{hp}{}
  \item{insurance_normloss}{The relative average loss payment per insured vehicle year. This value is normalized for all autos within a particular size classification (two-door small, station wagons, sports/speciality, etc...), and represents the average loss per car per year.}
  \item{insurance_symboling}{The degree to which the auto is more risky than its price indicates. Cars are initially assigned a risk factor symbol associated with its price. Then, if it is more risky (or less), this symbol is adjusted by moving it up (or down) the scale. Actuarians call this process "symboling". A value of +3 indicates that the auto is risky, -3 that it is probably pretty safe.}
  \item{length}{}
  \item{mpg_city}{}
  \item{mpg_highway}{}
  \item{price}{}
  \item{rpm_max}{}
  \item{weight}{}
  \item{wheelbase}{}
  \item{width}{}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Automobile}
}
\usage{
carprices
}
\description{
This data set consists of three types of entities: (a) the specification of
an auto in terms of various characteristics, (b) its assigned insurance risk
rating, (c) its normalized losses in use as compared to other cars.
}
\keyword{datasets}
