% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Modal Value
The mode, or modal value, is the most frequent value in a set of values.}
\usage{
modal(x, na.rm = FALSE)
}
\description{
Modal Value
The mode, or modal value, is the most frequent value in a set of values.
}
